please lint.
git-svn-id: file:///svn/unbound/trunk@3363 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/net_help.c
@@ -770,7 +770,7 @@ static lock_basic_t *ub_openssl_locks = NULL;
 static unsigned long
 ub_crypto_id_cb(void)
 {
-	return log_thread_get();
+	return (unsigned long)log_thread_get();
 }
 
 static void
